+++ data/src/main/java/com/github/slznvk/data/api/UserApiService.kt
@@ -21,10 +21,10 @@ interface UserApiService {
     suspend fun getAllUsers(): Response<List<User>>
 
     @GET("api/users/{id}")
-    suspend fun getUserById(@Path("id") id: Int): Response<User>
+    suspend fun getUserById(@Path("id") id: Long): Response<User>
 
     @GET("api/{userId}/jobs")
-    suspend fun getUserJobs(@Path("userId") userId: Int): Response<List<Job>>
+    suspend fun getUserJobs(@Path("userId") userId: Long): Response<List<Job>>
 
 //    @GET("api/{authorId}/wall")
 //    suspend fun getUserWall(@Path("authorId") authorId: Int): Response<List<Post>>
@@ -33,7 +33,7 @@ interface UserApiService {
     suspend fun saveJob(@Body job: Job): Response<Job>
 
     @DELETE("api/my/jobs")
-    suspend fun deleteJodById(@Path("id") id: Int)
+    suspend fun deleteJodById(@Path("id") id: Long)
 
 
     //AUTH
