+++ data/src/main/java/com/github/slznvk/data/entity/PostRemoteKeyEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class PostRemoteKeyEntity(
     @PrimaryKey
     val type: KeyType,
-    val key: Int
+    val key: Long
 ) {
     enum class KeyType {
         AFTER,
